Added UI Elements, Box Config file, additional Docker features, and updated the README
