--- box-node-express-skeleton-app/app-web/public/css/box.css
@@ -1,11 +1,0 @@
-.vertical-center {
-  min-height: 100%;  /* Fallback for browsers do NOT support vh unit */
-  min-height: 100vh; /* These two lines are counted as one :-)       */
-
-  display: flex;
-  align-items: center;
-}
-
-.boxPreviewSelector { 
-    height: 600px;
-}
