--- box-node-express-skeleton-app/app-web/views/pages/box/file-preview.ejs
@@ -1,20 +1,0 @@
-<% include ../../partials/common/header %>
-<div class="boxPreviewSelector"  id='boxPreview'></div>
-<a href="/files/<%= fileId %>">
-    Return
-</a>
-<script src="https://cdn01.boxcdn.net/platform/preview/0.126.0/en-US/preview.js"></script>
-<script type="text/javascript">
-    var accessToken = "<%= accessToken %>"; 
-    var fileId = "<%= fileId %>"; 
-    var boxPreview = new Box.Preview();
-
-    boxPreview.show(fileId, accessToken, {
-        container: '#boxPreview',
-        logoUrl: 'https://s-media-cache-ak0.pinimg.com/736x/a3/5d/a1/a35da1fb04988aa038974035501df5a0.jpg',
-        showDownload: true,
-        showAnnotations: true
-    });
-
-</script>
-<% include ../../partials/common/footer %>
