--- box-node-express-skeleton-app/app-web/views/pages/index.ejs
@@ -1,4 +1,0 @@
-<% include ../partials/common/header %>
-<h1>Welcome to Box!</h1>
-<a href="./about">About</a>
-<% include ../partials/common/footer %>
