--- box-node-express-skeleton-app/app-web/views/partials/box/folders.ejs
@@ -1,31 +1,0 @@
-<div id="folders" class="col-md-3">
-  <div class="row">
-    <div class="col-md-12">
-      <h4>Folders</h4>
-    </div>
-  </div>
-  <hr>
-  <form method="post" action="/user/<%= currentFolder %>/folder"> 
-    <div class="form-group">
-    <label for="folderName">Folder Name:</label> 
-                
-    <input type="text" name="folderName" placeholder="Folder name..." />                    
-    <button type='submit' class="btn btn-default">Add Folder</button>
-    </div>
-  </form>
-  <div id="foldersUi">
-    <% if(folders.length === 0) { %>
-      <p id="noFolders" class="warning text-center">No subfolders found.</p>
-    <% } %>
-    <ul id="folderGroup" class="list-group">
-    <% folders.forEach(function(folder) { %>
-      <li id="<%= folder.id %>" class="list-group-item">
-        <span class="folder" style="cursor: pointer;" data-id="<%= folder.id %>">
-          <i class="fa fa-folder fa-2x fa-fw" aria-hidden="true"></i>
-          <span class="folder-name"><a href="/user/<%= folder.id %>"><%= folder.name %></a></span> 
-        </span>
-      </li>
-    <% }); %>
-    </ul>
-  </div>
-</div>
