--- box-node-express-skeleton-app/app-web/views/partials/common/footer.ejs
@@ -1,5 +1,0 @@
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.2/jquery.min.js"></script>
-<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-<script src="https://cdn01.boxcdn.net/platform/preview/0.126.0/en-US/preview.js"></script>
-</body>
-</html>
