--- box-node-express-skeleton-app/app-web/views/partials/common/header.ejs
@@ -1,16 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <% if(locals.title) { %>
-    <title><%- title %></title>
-  <% } else { %>
-    <title>Box Skeleton App</title>
-  <% } %>
-  <link rel="stylesheet" href="/css/box.css">
-  <link rel="stylesheet" href="https://cdn01.boxcdn.net/platform/preview/0.126.0/en-US/preview.css" />
-  <link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" rel="stylesheet">
-  <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.6.3/css/font-awesome.min.css" rel="stylesheet">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-</head>
-<body>
