+++ box-node-express-skeleton-app/app/.eslintrc
