+++ box-node-express-skeleton-app/app/app-web/controllers/indexController.js
