+++ box-node-express-skeleton-app/app/app-web/views/pages/content-picker.ejs
@@ -1,0 +1,16 @@
+<% include ../partials/common/header %>
+<% include ../partials/common/display-header %>
+<% include ../partials/box/box-ui-selectors %>
+<div id="box-picker" class="container-fluid">
+</div>
+<% include ../partials/box/box-ui-scripts %>
+<script type="text/javascript">
+  var accessToken = "<%= accessToken %>"; 
+  var folderId = "<%= rootFolder %>";
+  var filePicker = new Box.FilePicker();
+  filePicker.show(folderId, accessToken, {
+    container: '#box-picker'
+});
+</script>
+
+<% include ../partials/common/footer %>
