+++ box-node-express-skeleton-app/app/app-web/views/partials/box/box-ui-scripts.ejs
@@ -1,0 +1,5 @@
+<script src="https://cdn01.boxcdn.net/platform/preview/1.1.1/en-US/preview.js"></script>
+<script src="https://cdn01.boxcdn.net/platform/elements/1.0.2/en-US/explorer.js"></script>
+<script src="https://cdn01.boxcdn.net/platform/elements/1.0.2/en-US/picker.js"></script>
+<script src="https://cdn01.boxcdn.net/platform/elements/1.0.2/en-US/uploader.js"></script>
+<script src="https://cdn01.boxcdn.net/platform/elements/1.0.2/en-US/tree.js"></script>
