+++ box-node-express-skeleton-app/app/app-web/views/partials/box/box-ui-selectors.ejs
@@ -1,0 +1,22 @@
+<nav class="navbar navbar-default">
+  <div class="container-fluid">
+    <div class="navbar-header">
+      <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
+        <span class="sr-only">Toggle navigation</span>
+        <span class="icon-bar"></span>
+        <span class="icon-bar"></span>
+        <span class="icon-bar"></span>
+      </button>
+      <a class="navbar-brand" href="#">Brand</a>
+    </div>
+
+    <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
+      <ul class="nav navbar-nav">
+        <li><a href="/user">Content Explorer</a></li>
+        <li><a href="/content-picker">Content Picker</a></li>
+        <li><a href="/content-uploader">Content Uploader</a></li>
+        <li><a href="/content-tree">Content Tree</a></li>
+      </ul>
+    </div>
+  </div>
+</nav>
