+++ box-node-express-skeleton-app/app/app-web/views/partials/common/display-header.ejs
@@ -1,0 +1,19 @@
+<div id="box-header" class="container-fluid">
+    <div class="jumbotron">
+      <h1>
+      <%= title %>
+      </h1>
+      <div class="container-fluid">
+        <div class="row">
+          <div class="col-md-1">
+            <img class="img-responsive" src="<%= user.picture %>" />
+          </div>
+          <div class="col-md-11">
+            <h2> Welcome
+            <%= user.nickname %>! 
+            </h2>
+          </div>
+        </div>
+      </div>
+    </div>
+</div>
