+++ box-node-express-skeleton-app/app/app-web/views/partials/common/footer.ejs
@@ -1,0 +1,4 @@
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.2/jquery.min.js"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+</body>
+</html>
