+++ box-node-express-skeleton-app/app/app-web/views/partials/common/header.ejs
@@ -1,0 +1,19 @@
+<!doctype html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <% if(locals.title) { %>
+    <title><%- title %></title>
+  <% } else { %>
+    <title>Box Skeleton App</title>
+  <% } %>
+  <link rel="stylesheet" href="/css/box.css">
+  <link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" rel="stylesheet">
+  <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.6.3/css/font-awesome.min.css" rel="stylesheet">
+  <link rel="stylesheet" href="https://cdn01.boxcdn.net/platform/preview/1.1.1/en-US/preview.css" />
+  <link rel="stylesheet" href="https://cdn01.boxcdn.net/platform/elements/1.0.2/en-US/explorer.css" />
+  <link rel="stylesheet" href="https://cdn01.boxcdn.net/platform/elements/1.0.2/en-US/picker.css" />
+  <link rel="stylesheet" href="https://cdn01.boxcdn.net/platform/elements/1.0.2/en-US/tree.css" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+</head>
+<body>
