+++ box-node-express-skeleton-app/app/gulpfile.js
@@ -1,0 +1,15 @@
+'use strict';
+const gulp = require('gulp'),
+	nodemon = require('gulp-nodemon'),
+	eslint = require('gulp-eslint');
+
+gulp.task('lint', () => {
+	return gulp.src(['**/*.js', '!node_modules/**'])
+		.pipe(eslint())
+		.pipe(eslint.format())
+		.pipe(eslint.failAfterError());
+});
+
+gulp.task('default', [
+	'lint'
+]);
