+++ box-node-express-skeleton-app/app/processes.json
@@ -1,0 +1,16 @@
+{
+    "apps": [
+        {
+            "name": "box-skeleton-app",
+            "script": "./bin/www",
+            "merge_logs": true,
+            "max_restarts": 20,
+            "instances": 4,
+            "max_memory_restart": "200M",
+            "env": {
+                "PORT": 3000,
+                "NODE_ENV": "production"
+            }
+        }
+    ]
+}
