+++ box-node-express-skeleton-app/docker-compose.yml
@@ -5,3 +5,12 @@ services:
       command: redis-server --requirepass securepassword
       ports:
         - '6379:6379'
+  app:
+    image: keymetrics/pm2-docker-alpine
+    volumes:
+      - "./app:/app"
+    ports:
+      - "3000:3000"
+    links:
+      - "redis"
+    command: "pm2 start ./processes.json --no-daemon"
